优化控件颜色

+++ app/src/main/java/com/github/umer0586/droidpad/ui/screens/controlpadbuilderscreen/ControlPadBuilderScreen.kt
@@ -299,7 +299,7 @@ fun ControlPadBuilderScreenContent(
         BoxWithConstraints(
             Modifier
                 .fillMaxSize()
-                .background(Color(controlPad.backgroundColor.toULong()))
+                .background(Color(controlPad.backgroundColor))
                 .padding(innerPadding)
                 .clipToBounds()
                 .clickable(
